mudei tudo de int pra long long

+++ algoritmos1/tp1/include/grafo.hpp
@@ -7,38 +7,38 @@
 
 class Grafo {
     public:
-        Grafo(int n, int m);
+        Grafo(long long int n, long long int m);
         ~Grafo();
         
         void LerStdIn();
-        int MenorCaminhoEntre1eN();
-        std::set<int> RuasEmCaminhoMinimo();
-        std::set<int> RuasCriticas();
+        long long int MenorCaminhoEntre1eN();
+        std::set<long long int> RuasEmCaminhoMinimo();
+        std::set<long long int> RuasCriticas();
         
     private:
         // _n corresponde ao número de vértices e _m ao número de arestas
-        int _n, _m;
+        long long int _n, _m;
 
         // Lista de adjacência será utilizada para caminhamento no grafo
-        std::vector<std::vector<std::pair<int, int>>> _listaDeAdjacencia;
+        std::vector<std::vector<std::pair<long long int, long long int>>> _listaDeAdjacencia;
         
         // Armazenar as arestas originais
         struct Aresta {
-            int u, v, peso, indice;
+            long long int u, v, peso, indice;
         };
         std::vector<Aresta> _arestas;
 
         // Funcionalidades para armazenar o caminhamento do grafo
         void Dijkstra(
-            int origem,
-            std::vector<int>& distancias,
-            std::vector<std::vector<int>>& predecessores
+            long long int origem,
+            std::vector<long long int>& distancias,
+            std::vector<std::vector<long long int>>& predecessores
         );
         void EncontrarCaminhosMinimos(
-            std::vector<std::vector<int>>& predecessores,
-            int destino,
-            int atual,
-            std::vector<int>& caminho,
-            std::set<int>& arestasEmCaminhoMinimo
+            std::vector<std::vector<long long int>>& predecessores,
+            long long int destino,
+            long long int atual,
+            std::vector<long long int>& caminho,
+            std::set<long long int>& arestasEmCaminhoMinimo
         );
 };
