+++ algoritmos1/tp1/src/grafo.cpp
@@ -5,7 +5,7 @@
 #include <climits>
 #include <algorithm>
 
-Grafo::Grafo(int n, int m) {
+Grafo::Grafo(long long int n, long long int m) {
     // Inicializa número de vértices e arestas do grafo
     this->_n = n;
     this->_m = m;
@@ -15,8 +15,8 @@ Grafo::Grafo(int n, int m) {
 }
 
 void Grafo::LerStdIn() {
-    int u, v, c;
-    for (int i = 0; i < this->_m; i++) {
+    long long int u, v, c;
+    for (long long int i = 0; i < this->_m; i++) {
         std::cin >> u >> v >> c;
         u--; v--; // Ajuste de input 1-indexado
         this->_listaDeAdjacencia[u].push_back({v, c});
@@ -26,17 +26,17 @@ void Grafo::LerStdIn() {
     }
 }
 
-void Grafo::Dijkstra(int origem, std::vector<int>& distancias, std::vector<std::vector<int>>& predecessores) {
+void Grafo::Dijkstra(long long int origem, std::vector<long long int>& distancias, std::vector<std::vector<long long int>>& predecessores) {
     // Inicializa distâncias e visitados
     distancias.assign(this->_n, INT_MAX);
     std::vector<bool> visitados(this->_n, false);
-    predecessores.assign(this->_n, std::vector<int>());
+    predecessores.assign(this->_n, std::vector<long long int>());
     
     // Criação de um min heap para os custos
     std::priority_queue<
-        std::pair<int, int>, 
-        std::vector<std::pair<int, int>>, 
-        std::greater<std::pair<int, int>>
+        std::pair<long long int, long long int>, 
+        std::vector<std::pair<long long int, long long int>>, 
+        std::greater<std::pair<long long int, long long int>>
     > minHeap;
 
     // Inicialização do heap com o vértice inicial
@@ -45,11 +45,11 @@ void Grafo::Dijkstra(int origem, std::vector<int>& distancias, std::vector<std::
 
     while (!minHeap.empty()) {
         // Pega o vértice com menor custo
-        std::pair<int, int> v = minHeap.top();
+        std::pair<long long int, long long int> v = minHeap.top();
         minHeap.pop();
 
-        int distanciaV = v.first;
-        int indiceV = v.second;
+        long long int distanciaV = v.first;
+        long long int indiceV = v.second;
 
         // Verifica se ele já foi visitado
         if (visitados[indiceV]) continue;
@@ -57,8 +57,8 @@ void Grafo::Dijkstra(int origem, std::vector<int>& distancias, std::vector<std::
 
         // Para cada vizinho dele, atualizamos, se necessário, o custo para chegar em outros vértices
         for (auto vizinho : _listaDeAdjacencia[indiceV]) {
-            int u = vizinho.first;
-            int peso = vizinho.second;
+            long long int u = vizinho.first;
+            long long int peso = vizinho.second;
             
             if (distanciaV + peso < distancias[u]) {
                 distancias[u] = distanciaV + peso;
@@ -73,18 +73,18 @@ void Grafo::Dijkstra(int origem, std::vector<int>& distancias, std::vector<std::
     }
 }
 
-int Grafo::MenorCaminhoEntre1eN() {
-    std::vector<int> distancias;
-    std::vector<std::vector<int>> predecessores;
+long long int Grafo::MenorCaminhoEntre1eN() {
+    std::vector<long long int> distancias;
+    std::vector<std::vector<long long int>> predecessores;
     Dijkstra(0, distancias, predecessores);
     return distancias[this->_n - 1];
 }
 
-void Grafo::EncontrarCaminhosMinimos(std::vector<std::vector<int>>& predecessores, int destino, int atual, std::vector<int>& caminho, std::set<int>& arestasEmCaminhoMinimo) {
+void Grafo::EncontrarCaminhosMinimos(std::vector<std::vector<long long int>>& predecessores, long long int destino, long long int atual, std::vector<long long int>& caminho, std::set<long long int>& arestasEmCaminhoMinimo) {
     if (atual == 0) { // Chegamos à origem
         for (size_t i = 0; i < caminho.size() - 1; i++) {
-            int u = caminho[i];
-            int v = caminho[i + 1];
+            long long int u = caminho[i];
+            long long int v = caminho[i + 1];
             
             // Encontrar o índice da aresta entre u e v
             for (const auto& aresta : _arestas) {
@@ -97,34 +97,34 @@ void Grafo::EncontrarCaminhosMinimos(std::vector<std::vector<int>>& predecessore
         return;
     }
     
-    for (int pred : predecessores[atual]) {
+    for (long long int pred : predecessores[atual]) {
         caminho.push_back(pred);
         EncontrarCaminhosMinimos(predecessores, destino, pred, caminho, arestasEmCaminhoMinimo);
         caminho.pop_back();
     }
 }
 
-std::set<int> Grafo::RuasEmCaminhoMinimo() {
-    std::vector<int> distancias;
-    std::vector<std::vector<int>> predecessores;
+std::set<long long int> Grafo::RuasEmCaminhoMinimo() {
+    std::vector<long long int> distancias;
+    std::vector<std::vector<long long int>> predecessores;
     Dijkstra(0, distancias, predecessores);
     
-    std::set<int> arestasEmCaminhoMinimo;
-    std::vector<int> caminho = {this->_n - 1}; // Começamos do destino
+    std::set<long long int> arestasEmCaminhoMinimo;
+    std::vector<long long int> caminho = {this->_n - 1}; // Começamos do destino
     EncontrarCaminhosMinimos(predecessores, this->_n - 1, this->_n - 1, caminho, arestasEmCaminhoMinimo);
     
     return arestasEmCaminhoMinimo;
 }
 
-std::set<int> Grafo::RuasCriticas() {
-    std::set<int> ruasCriticas;
-    std::set<int> ruasEmCaminhoMinimo = RuasEmCaminhoMinimo();
-    int distanciaOriginal = MenorCaminhoEntre1eN();
+std::set<long long int> Grafo::RuasCriticas() {
+    std::set<long long int> ruasCriticas;
+    std::set<long long int> ruasEmCaminhoMinimo = RuasEmCaminhoMinimo();
+    long long int distanciaOriginal = MenorCaminhoEntre1eN();
     
     // Para cada aresta em um caminho mínimo, verificamos se ela é crítica
-    for (int indiceAresta : ruasEmCaminhoMinimo) {
+    for (long long int indiceAresta : ruasEmCaminhoMinimo) {
         // Encontrar a aresta correspondente
-        int u = -1, v = -1;
+        long long int u = -1, v = -1;
         for (size_t i = 0; i < _arestas.size(); i++) {
             if (_arestas[i].indice == indiceAresta) {
                 u = _arestas[i].u;
@@ -141,15 +141,15 @@ std::set<int> Grafo::RuasCriticas() {
                 [u](const auto& p) { return p.first == u; });
             
             if (it_u != _listaDeAdjacencia[u].end() && it_v != _listaDeAdjacencia[v].end()) {
-                int peso_u = it_u->second;
-                int peso_v = it_v->second;
+                long long int peso_u = it_u->second;
+                long long int peso_v = it_v->second;
                 
                 _listaDeAdjacencia[u].erase(it_u);
                 _listaDeAdjacencia[v].erase(it_v);
                 
                 // Calcular nova distância mínima
-                std::vector<int> novasDistancias;
-                std::vector<std::vector<int>> novosPredecessores;
+                std::vector<long long int> novasDistancias;
+                std::vector<std::vector<long long int>> novosPredecessores;
                 Dijkstra(0, novasDistancias, novosPredecessores);
                 
                 // Se a distância aumentou ou ficou infinita, a aresta é crítica
